[InstCombine] Treat min/max as select in binop fold part 2.
This is a generalization to the code added in #65978, attempting to handle more
with predicates and min/max that match. It was originally trying to use
matchDecomposedSelectPattern but I replaced that with just checking the
predicates and SPF match. The opposite conditions are also handled, which just
inverts the order of the E/F variables.

TODO: Add some proofs.