[AArch64] Remove AdjustCost from getCastInstrCost.
This adds codesize costs for getCastInstrCost by removing the AdjustCost binary
function, allowing the existing throughput costs to be added for Codesize,
SizeAndLat and Latency costs. The latency costs are probably not very accurate,
but will be closer and should probably eventually come from the scheduling
model if we can prove out that works.